+++ hutool-system/src/test/java/cn/hutool/system/OshiTest.java
@@ -1,9 +1,9 @@
 package cn.hutool.system;
 
+import cn.hutool.system.oshi.CpuInfo;
+import cn.hutool.system.oshi.OshiUtil;
 import org.junit.Assert;
 import org.junit.Test;
-
-import cn.hutool.system.oshi.OshiUtil;
 
 public class OshiTest {
 	
@@ -11,5 +11,11 @@ public class OshiTest {
 	public void getMemeryTest() {
 		long total = OshiUtil.getMemory().getTotal();
 		Assert.assertTrue(total > 0);
+	}
+
+	@Test
+	public void getCupInfo() {
+		CpuInfo cpuInfo = OshiUtil.getCpuInfo();
+		Assert.assertNotNull(cpuInfo);
 	}
 }